feat: all things

+++ src/album/album.service.ts
@@ -1,8 +1,14 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common';
+import { Logger } from '@nestjs/common/services';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Favorite } from 'src/favorite/entities/favorite.entity';
 import { Track } from 'src/track/entities/track.entity';
 import { Repository } from 'typeorm';
+import { validate } from 'uuid';
 import { CreateUpdateAlbumDto } from './dto/create-update-album.dto';
 import { Album } from './entities/album.entity';
 
@@ -23,23 +29,33 @@ export class AlbumService {
   }
 
   async create(AlbumDto: CreateUpdateAlbumDto): Promise<Album> {
-    const createAlbum = await this.albumRepository.create(AlbumDto);
+    const createAlbum = this.albumRepository.create(AlbumDto);
     return await this.albumRepository.save(createAlbum);
   }
 
   async findOne(id: string): Promise<Album> {
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
+
     const album: Album | null = await this.albumRepository.findOne({
       where: { id },
     });
-
     if (!album) {
-      throw new NotFoundException('User not found');
+      throw new NotFoundException('Album not found');
     }
 
     return album;
   }
 
   async updateOne(albumDto: CreateUpdateAlbumDto, id: string): Promise<Album> {
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
     let album: Album | null = await this.albumRepository.findOne({
       where: { id },
     });
@@ -53,16 +69,22 @@ export class AlbumService {
     return album;
   }
 
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  async updateArtistId(id: string) {}
-
   async delete(id: string) {
-    await this.albumRepository.delete(id);
-    const track = (
-      await this.trackRepository.find({ where: { albumId: id } })
-    )[0];
-    track.albumId = null;
-    await this.trackRepository.save(track);
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
+    const album = await this.albumRepository.delete(id);
+    if (album.affected === 0) throw new NotFoundException('Album is not found');
+    const track = await this.trackRepository.findOne({
+      where: { albumId: id },
+    });
+
+    if (track) {
+      track.albumId = null;
+      await this.trackRepository.save(track);
+    }
     const favorite = (
       await this.favoriteRepository.find({ relations: { albums: true } })
     )[0];
