+++ src/artist/artist.service.ts
@@ -1,9 +1,15 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  BadRequestException,
+} from '@nestjs/common';
+import { Logger } from '@nestjs/common/services';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Album } from 'src/album/entities/album.entity';
 import { Favorite } from 'src/favorite/entities/favorite.entity';
 import { Track } from 'src/track/entities/track.entity';
 import { Repository } from 'typeorm';
+import { validate } from 'uuid';
 import { CreateUpdateArtistDto } from './dto/create-update-track.dto';
 import { Artist } from './entities/artist.entity';
 
@@ -31,6 +37,11 @@ export class ArtistService {
   }
 
   async findOne(id: string): Promise<Artist> {
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
     const artist: Artist | null = await this.artistRepository.findOne({
       where: { id },
     });
@@ -46,6 +57,11 @@ export class ArtistService {
     artistDto: CreateUpdateArtistDto,
     id: string,
   ): Promise<Artist> {
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
     let artist: Artist | null = await this.artistRepository.findOne({
       where: { id },
     });
@@ -60,22 +76,35 @@ export class ArtistService {
   }
 
   async delete(id: string) {
-    await this.artistRepository.delete(id);
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
+    const artist = await this.artistRepository.delete(id);
+    if (artist.affected === 0)
+      throw new NotFoundException('Artist is not found');
     const favorite = (
       await this.favoriteRepository.find({ relations: { artists: true } })
     )[0];
     const indexFav = favorite.artists.findIndex((artist) => artist.id === id);
     favorite.artists.splice(indexFav, 1);
     await this.favoriteRepository.save(favorite);
-    const track = (
-      await this.trackRepository.find({ where: { artistId: id } })
-    )[0];
-    track.artistId = null;
-    await this.trackRepository.save(track);
-    const album = (
-      await this.albumRepository.find({ where: { artistId: id } })
-    )[0];
-    album.artistId = null;
-    await this.albumRepository.save(album);
+    const track = await this.trackRepository.findOne({
+      where: { artistId: id },
+    });
+
+    if (track) {
+      track.artistId = null;
+      await this.trackRepository.save(track);
+    }
+
+    const album = await this.albumRepository.findOne({
+      where: { artistId: id },
+    });
+    if (album) {
+      album.artistId = null;
+      await this.albumRepository.save(album);
+    }
   }
 }
