+++ src/track/track.service.ts
@@ -1,7 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Favorite } from 'src/favorite/entities/favorite.entity';
 import { Repository } from 'typeorm';
+import { validate } from 'uuid';
 import { CreateUpdateTrackDto } from './dto/create-update-track.dto';
 import { Track } from './entities/track.entity';
 
@@ -25,6 +30,11 @@ export class TrackService {
   }
 
   async findOne(id: string): Promise<Track> {
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
     const track: Track | null = await this.trackRepository.findOne({
       where: { id },
     });
@@ -37,6 +47,11 @@ export class TrackService {
   }
 
   async updateOne(trackDto: CreateUpdateTrackDto, id: string): Promise<Track> {
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
     let track: Track | null = await this.trackRepository.findOne({
       where: { id },
     });
@@ -50,22 +65,15 @@ export class TrackService {
     return track;
   }
 
-  async updateArtistId(id: string) {
-    const track: Track | null = await this.trackRepository.findOne({
-      where: { artistId: id },
-    });
-    await this.trackRepository.save(track);
-  }
-
-  async updateAlbumId(id: string) {
-    const track: Track | null = await this.trackRepository.findOne({
-      where: { albumId: id },
-    });
-    await this.trackRepository.save(track);
-  }
-
   async delete(id: string) {
-    await this.trackRepository.delete(id);
+    const compareId = validate(id);
+
+    if (!compareId) {
+      throw new BadRequestException('Id is not valid');
+    }
+    const track = await this.trackRepository.delete(id);
+    if (track.affected === 0)
+      throw new NotFoundException('Track is not founded');
     const favorite = (
       await this.favoriteRepository.find({ relations: { tracks: true } })
     )[0];
